Renamed GetMessageById to RetrieveMessageById
