+++ MessagingApp.Application/ApplicationDependencyInjection.cs
@@ -7,8 +7,8 @@ using MessagingApp.Application.Users.Commands.CreateFriendRequest;
 using MessagingApp.Application.Users.Commands.CreateUser;
 using MessagingApp.Application.Users.Commands.SendMessage;
 using MessagingApp.Application.Users.Queries.AuthenticateUser;
-using MessagingApp.Application.Users.Queries.GetMessageById;
 using MessagingApp.Application.Users.Queries.RetrieveConversation;
+using MessagingApp.Application.Users.Queries.RetrieveMessageById;
 using MessagingApp.Application.Users.Queries.RetrieveUser;
 
 // ReSharper disable once CheckNamespace
@@ -31,7 +31,7 @@ public static class ApplicationDependencyInjection
         
         // Register messaging handlers
         services.AddTransient<IHandler<SendMessageCommand, SendMessageResponse>, SendMessageHandler>();
-        services.AddTransient<IHandler<GetMessageByIdQuery, GetMessageResponse>, GetMessageByIdHandler>();
+        services.AddTransient<IHandler<RetrieveMessageByIdQuery, GetMessageResponse>, RetrieveMessageByIdHandler>();
         services
             .AddTransient<IHandler<RetrieveConversationQuery, RetrieveConversationResponse>,
                 RetrieveConversationHandler>();
