+++ MessagingApp.Application/Users/Queries/RetrieveMessageById/RetrieveMessageByIdHandler.cs
@@ -4,18 +4,18 @@ using MessagingApp.Application.Common.Exceptions;
 using MessagingApp.Application.Common.Interfaces.Mediator;
 using MessagingApp.Application.Common.Interfaces.Repositories;
 
-namespace MessagingApp.Application.Users.Queries.GetMessageById;
+namespace MessagingApp.Application.Users.Queries.RetrieveMessageById;
 
-public class GetMessageByIdHandler : IHandler<GetMessageByIdQuery, GetMessageResponse>
+public class RetrieveMessageByIdHandler : IHandler<RetrieveMessageByIdQuery, GetMessageResponse>
 {
     private readonly IUserRepository _userRepository;
 
-    public GetMessageByIdHandler(IUserRepository userRepository)
+    public RetrieveMessageByIdHandler(IUserRepository userRepository)
     {
         _userRepository = userRepository;
     }
 
-    public async Task<Result<GetMessageResponse>> Handle(GetMessageByIdQuery req)
+    public async Task<Result<GetMessageResponse>> Handle(RetrieveMessageByIdQuery req)
     {
         try
         {
