+++ MessagingApp.Application/Users/Queries/RetrieveMessageById/RetrieveMessageByIdQuery.cs
@@ -1,14 +1,14 @@
 ﻿using MessagingApp.Application.Common.Contracts;
 using MessagingApp.Application.Common.Interfaces.Mediator;
 
-namespace MessagingApp.Application.Users.Queries.GetMessageById;
+namespace MessagingApp.Application.Users.Queries.RetrieveMessageById;
 
-public class GetMessageByIdQuery : IRequest<GetMessageResponse>
+public class RetrieveMessageByIdQuery : IRequest<GetMessageResponse>
 {
     public Guid MessageId { get; set; }
     public Guid RequestingUserId { get; set; }
 
-    public GetMessageByIdQuery(Guid messageId, Guid requestingUser)
+    public RetrieveMessageByIdQuery(Guid messageId, Guid requestingUser)
     {
         MessageId = messageId;
         RequestingUserId = requestingUser;
